Separate files and cleanup
