+++ aries-site/src/layouts/content/Example/index.js
@@ -1,0 +1,6 @@
+export * from './BrowserWrapper';
+export * from './Example';
+export * from './ExampleControls';
+export * from './ExampleResources';
+export * from './ResponsiveControls';
+export * from './syntax';
