+++ aries-site/src/layouts/content/index.js
@@ -2,10 +2,8 @@ export * from './ButtonRow';
 export * from './ColorRow';
 export * from './ContentSection';
 export * from './Example';
-export * from './ExampleControls';
 export * from './FeedbackSection';
 export * from './Head';
 export * from './PageIntro';
-export * from './ResponsiveControls';
 export * from './Subsection';
 export * from './UsageExample';
